+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 ## [1.0.3] - 2020-07-08
 
-- Async validators (only FormControl for now...)
+### New Validators
+- **Async Validators** (only FormControl for now...)
 - Controls have now three different states: **VALID**, **INVALID**, **PENDING** 
   (this last one was specially included due to async validators, the control is **PENDING** 
   until validator completes)
 - Added documentations about *Validator.pattern*
 - Added *Async Validator's* example in example application */example/main.dart* 
+
+### New Models
+- **FormArray** (aggregates the values of each child FormControl into an array)
+
+### New Reactive Widgets:
+- ReactiveFormArray
+- ReactiveCheckboxListTile
+- ReactiveSwitchListTile
+- ReactiveRadioListTile
+- ReactiveStatusListenableBuilder
 
 ## [1.0.2] - 2020-07-05